+++ plugins/jsonparser/changelog.md
@@ -1,4 +1,4 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- chore: version update core to 1.2.17 and framework to 1.1.19
+- chore: version update core to 1.2.18 and framework to 1.1.21